Lots of Work
Finished major reorganizing of the second version. Changed js onload
from html to javascript. Updated the ppt. Really worked on the
navigation for the second version. About to transfer the navigation to
the first version. One of the next steps is to clean up the css: remove
useless css, combine repeated “headers” (forgot what they’re called),
and add some comments to the code that I have to “relearn” every time.